# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API do controle de homologação de projetos solares da ConectSol.\n\nAutenticação: POST /api/auth/login (e-mail e senha). Use o accessToken como Bearer. Não há auto-cadastro: as contas são criadas por ADMINISTRADOR ou GESTOR em POST /api/usuarios. Não há logout no servidor — a API é stateless e o cliente descarta os tokens.\n\nMudanças de status são feitas por endpoints de ação (ex.: POST /api/pendencias/{id}/resolver), nunca por PUT: é o que garante a automação entre etapas e a auditoria.",
    "title" : "SolarSync API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginSenhaRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "Tokens emitidos"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "Credenciais inválidas"
          }
        },
        "summary" : "Login com e-mail e senha",
        "tags" : [ "Autenticação" ]
      }
    },
    "/api/auth/refresh" : {
      "post" : {
        "description" : "O usuário é relido no banco: quem foi desativado não renova. Não há logout no servidor — a API é stateless, o cliente descarta os tokens.",
        "operationId" : "renovar",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "Refresh token inválido, expirado ou de usuário inativo"
          }
        },
        "summary" : "Troca o refresh token por um novo par de tokens",
        "tags" : [ "Autenticação" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "LoginSenhaRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "senha" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "email", "senha" ]
      },
...
      "RefreshRequest" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "refreshToken" ]
      },
...
      "TokenResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiraEmSegundos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "tipo" : {
            "type" : "string"
          },
          "usuario" : {
            "$ref" : "#/components/schemas/UsuarioLogadoResponse"
          }
        }
      },
...
      "UsuarioLogadoResponse" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nome" : {
            "type" : "string"
          },
          "papeis" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "ADMINISTRADOR", "GESTOR", "ANALISTA" ]
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}